missing int casting

+++ core/scheduler.py
@@ -447,9 +447,9 @@ class SlurmScheduler(Scheduler):
         walltime = self.config.param(job_name_prefix, 'cluster_walltime')
         # force the DD-HH:MM[:00] format to HH:MM[:00]
         time = utils.time_to_datetime(walltime)
-        sec = time.seconds % 60
-        minutes = ((time.seconds - sec) / 60) % 60
-        hours = (time.seconds - sec - 60 * minutes) / 3600 + time.days * 24
+        sec = int(time.seconds % 60)
+        minutes = int(((time.seconds - sec) / 60) % 60)
+        hours = int((time.seconds - sec - 60 * minutes) / 3600 + time.days * 24)
         return '--time={:02d}:{:02d}:{:02d}'.format(hours, minutes, sec)
 
     def memory(self, job_name_prefix):
